requirements.txt
asgiref==3.6.0
colorama==0.4.6
Django==3.0.6
Pillow==9.5.0
pypng==0.20220715.0
pytz==2023.3
qrcode==7.4.2
sqlparse==0.4.3
typing-extensions==4.5.0
